Silence most warnings
Suppress and fix many clang, GCC and MSVC warnings

"cause the night is crap and full of warnings"

+++ chromium/base/allocator/partition_allocator/src/partition_alloc/page_allocator_constants.h
@@ -160,7 +160,7 @@ PageAllocationGranularity() {
   }
   return size;
 #else
-  return 1 << PageAllocationGranularityShift();
+  return size_t(1) << PageAllocationGranularityShift();
 #endif
 }
 
@@ -195,7 +195,7 @@ SystemPageSize() {
   // on a non-constexpr.
   return PageAllocationGranularity();
 #else
-  return 1 << SystemPageShift();
+  return size_t(1) << SystemPageShift();
 #endif
 }
